backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
python-multipart==0.0.6
email-validator==2.1.1

# Environment Variables
python-dotenv==1.0.0

# External APIs - Specific versions for compatibility
cohere==4.48
qdrant-client==1.7.0
openai==1.10.0
httpx==0.26.0

# Rate Limiting
slowapi==0.1.9

# Utilities
beautifulsoup4==4.12.2
